Fixed line processing parallelism
Since java.nio.file.Files#list(Path) method returns a stream of unknown size, it's not possible for Stream Framework to effectively parallelize the further pipeline.
So the provided change is that a directory files list is fully retrieved and then converted to a stream (which is both java.util.Spliterator.SIZED and SUBSIZED due to the implementation of a list spliterator) thus allowing an effective parallelization of lines reading and processing pipeline.

Fixed parallelism verified using VisualVM profiling.