Improved line processing parallelism when using JDK 9+
Starting from JDK 9, java.nio.file.Files#lines(Path) method returns a stream with better parallelization capabilities (if a file is in local file system and its size is determinable), allowing to process groups of lines from the same file in different threads.
But this parallelization possibility is negated by using flat-mapping, which leads to retrieving items of an upstream sequentially, so every file is entirely processed by one thread. It is tolerable if the files are approximately the same size, but unacceptable if any file is significantly larger than others.
Replaced flat-mapping of lines streams with a balanced tree of concatenated streams to achieve better parallelism.

Better parallelism verified using VisualVM profiling.